+++ Application/UI/mvc-frontend/ImageRecognition/ImageRecognitionManager.cs
@@ -1,4 +1,9 @@
-﻿using Amazon.DynamoDBv2;
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
+using Amazon.DynamoDBv2;
 using Amazon.DynamoDBv2.DataModel;
 using Amazon.DynamoDBv2.DocumentModel;
 using Amazon.S3;
@@ -6,41 +11,39 @@ using Amazon.S3.Model;
 using Amazon.StepFunctions;
 using ImageRecognition.Frontend.Models;
 using Microsoft.Extensions.Options;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace ImageRecognition.Frontend
 {
     public class ImageRecognitionManager
     {
-        AppOptions _appOptions;
-        IAmazonDynamoDB _ddbClient;
-        IAmazonS3 _s3Client;
-        IAmazonStepFunctions _stepClient;
-        DynamoDBContext _ddbContext;
+        private readonly AppOptions _appOptions;
+        private readonly IAmazonDynamoDB _ddbClient;
+        private readonly DynamoDBContext _ddbContext;
+        private readonly IAmazonS3 _s3Client;
+        private IAmazonStepFunctions _stepClient;
 
-        public ImageRecognitionManager(IOptions<AppOptions> appOptions, IAmazonDynamoDB dbClient, IAmazonS3 s3Client, IAmazonStepFunctions stepClient)
+        public ImageRecognitionManager(IOptions<AppOptions> appOptions, IAmazonDynamoDB dbClient, IAmazonS3 s3Client,
+            IAmazonStepFunctions stepClient)
         {
             _appOptions = appOptions.Value;
             _ddbClient = dbClient;
             _s3Client = s3Client;
             _stepClient = stepClient;
-            _ddbContext = new DynamoDBContext(this._ddbClient);
+            _ddbContext = new DynamoDBContext(_ddbClient);
         }
+
         // Add photo to Album.
         // insert into photo dynamo table as pending
         // upload to S3
         public async Task<Photo> AddPhoto(string albumId, string userId, string name, Stream stream)
         {
             var tempFile = Path.GetTempFileName();
-            try 
+            try
             {
-                var photoId = name + "-" + Guid.NewGuid().ToString();
+                var photoId = name + "-" + Guid.NewGuid();
 
-                var photo = new Photo { 
+                var photo = new Photo
+                {
                     AlbumId = albumId,
                     Bucket = _appOptions.PhotoStorageBucket,
                     CreatedDate = DateTime.UtcNow,
@@ -51,7 +54,7 @@ namespace ImageRecognition.Frontend
                     UploadTime = DateTime.UtcNow
                 };
 
-                await this._ddbContext.SaveAsync(photo).ConfigureAwait(false);
+                await _ddbContext.SaveAsync(photo).ConfigureAwait(false);
 
                 using (var fileStream = File.OpenWrite(tempFile))
                 {
@@ -60,23 +63,19 @@ namespace ImageRecognition.Frontend
 
                 var putRequest = new PutObjectRequest
                 {
-                    BucketName = this._appOptions.PhotoStorageBucket,
+                    BucketName = _appOptions.PhotoStorageBucket,
                     Key = $"private/uploads/{userId}/{photoId}",
                     FilePath = tempFile
                 };
 
-                await this._s3Client.PutObjectAsync(putRequest).ConfigureAwait(false);
+                await _s3Client.PutObjectAsync(putRequest).ConfigureAwait(false);
 
                 return photo;
             }
             finally
             {
-                if (File.Exists(tempFile))
-                {
-                    File.Delete(tempFile);
-                }
+                if (File.Exists(tempFile)) File.Delete(tempFile);
             }
-
         }
 
         // Get Step function status.
@@ -86,15 +85,16 @@ namespace ImageRecognition.Frontend
         //  insert entry into Album dynamo db table.
         public async Task<Album> CreateAlbum(string userId, string albumName)
         {
-            var album = new Album { 
-                AlbumId = albumName + "-" + Guid.NewGuid().ToString(),
+            var album = new Album
+            {
+                AlbumId = albumName + "-" + Guid.NewGuid(),
                 Name = albumName,
                 UserId = userId,
                 CreateDate = DateTime.UtcNow,
                 UpdatedDate = DateTime.UtcNow
             };
 
-            await this._ddbContext.SaveAsync(album).ConfigureAwait(false);
+            await _ddbContext.SaveAsync(album).ConfigureAwait(false);
 
             return album;
         }
@@ -102,7 +102,7 @@ namespace ImageRecognition.Frontend
         // Get Albums by user
         public async Task<IList<Album>> GetAlbums(string userId)
         {
-            var search = this._ddbContext.QueryAsync<Album>(userId);
+            var search = _ddbContext.QueryAsync<Album>(userId);
 
             return await search.GetRemainingAsync().ConfigureAwait(false);
         }
@@ -111,19 +111,19 @@ namespace ImageRecognition.Frontend
         // Get Album Details.
         public async Task<Album> GetAlbumDetails(string userId, string albumId)
         {
-            var albumQuery = 
-                this._ddbContext.QueryAsync<Album>(userId, QueryOperator.Equal, new[] { albumId });
+            var albumQuery =
+                _ddbContext.QueryAsync<Album>(userId, QueryOperator.Equal, new[] {albumId});
 
             var albums = await albumQuery.GetRemainingAsync().ConfigureAwait(false);
 
-            Album album = albums.FirstOrDefault();
+            var album = albums.FirstOrDefault();
 
             if (album != null)
             {
-                var photoQuery = this._ddbContext.QueryAsync<Photo>(albumId, new DynamoDBOperationConfig { IndexName = "albumID-uploadTime-index" });
+                var photoQuery = _ddbContext.QueryAsync<Photo>(albumId,
+                    new DynamoDBOperationConfig {IndexName = "albumID-uploadTime-index"});
                 album.Photos = await photoQuery.GetRemainingAsync().ConfigureAwait(false);
                 foreach (var photo in album.Photos)
-                {
                     if (photo.ProcessingStatus == ProcessingStatus.Succeeded)
                     {
                         photo.Thumbnail.Url = _s3Client.GetPreSignedURL(new GetPreSignedUrlRequest
@@ -139,20 +139,19 @@ namespace ImageRecognition.Frontend
                             Expires = DateTime.UtcNow.AddHours(1)
                         });
                     }
-                }
             }
+
             return album;
         }
 
 
         public async Task<Photo> GetPhotoDetails(string userId, string photoId)
         {
-            var search = this._ddbContext.QueryAsync<Photo>(photoId);
+            var search = _ddbContext.QueryAsync<Photo>(photoId);
 
             var photos = await search.GetRemainingAsync().ConfigureAwait(false);
 
             return photos.Where(p => p.Owner == userId).FirstOrDefault();
         }
-
     }
 }
